rename client file; update imports
